backend/requirements.txt
Django==5.0.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
drf-yasg==1.21.10
python-decouple==3.8
django-cors-headers==4.3.1
pytest==8.2.2
pytest-django==4.8.0
black==24.4.2
isort==5.13.2
flake8==7.0.0
